rename contrastive.py -> self_supervised.py
